pyproject.toml
[project]
name = "lang-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.4",
    "langchain-google-genai>=4.1.2",
    "langchain-ollama>=1.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
